Change how data is split in favour of sklearn func
This removes references to 'validation' sets, which are now just
training and test sets, where the training set is split further during
hyperparameter optimisation

Allows for training to consider a 'full' training set.

This is also now matches the methodology used by Catarina et al in
preparation for comparison with the benchmark results

See discussion in the issues linked below for the reasoning behind this.

Making this change also makes allows for cross-validation during
hyperparameter optimisation easier

Refs:
    - https://github.com/keras-team/keras/issues/1753

	modified:   astronet/t2/preprocess.py
	modified:   astronet/t2/tests/int/test_train.py
	modified:   astronet/t2/tests/unit/test_import.py
	modified:   astronet/t2/tests/unit/test_preprocess.py
	modified:   astronet/t2/tests/unit/test_utils.py
	modified:   astronet/t2/train.py
	modified:   astronet/t2/utils.py