Load parameters from hyperopt that gives min loss
Since the objective function with Optuna now in the direction of
'minimize', the loss will be written as a positive number, and we would
want the parameters related to the objective score that has the lowest
loss.

Previously it was the case that the objective function was in the
'maximise' direction of the negative of score[0], i.e. negative of the
loss in the direction of maximise.

This should make things a little more clearer to the user.

There may be changes required to results.json files to account for this
so that old results are not loaded.

	modified:   train.py