Add further debugging print statements
This shows that the discrepancy in logloss values is down to the fact
that model.evaluate calculates the loss in batches and then aggregates,
whereas for model.predict, whilst it may calculate the predictions in
batches for performance, the associated logloss is calculated in one go,
i.e. on all the test input data.

For consistency and reproducibility, model.evaluate is now being
produced with a batch size of length of the full test data

	modified:   astronet/t2/train.py