Change approach calculating loss of hyper model
Since it has been observed that model.evaluate on varying batch sizes
gives significantly different results, it might be best to just run a
prediction over the validation set to get an idea of the real
performance.

It is suspected this occurs since it is 'evaluating' the model per batch
size, and if there is a good or bad batch, this will skew the results,
whereas 'predictions' are per-sample (and just loaded into memory in
batches) which would imply a more robust calculation. An intuitive way of
thinking of this is that IF batch_size = X_test.shape[0], i.e the full
size of the 'validation set', then model.evaluate should ==
model.predict.

Due to OOM errors this can not be properly tested on this data, but it
was been verified as to be the case on other data.

	modified:   astronet/t2/opt/hypertrain.py