Reverting to eager mode and single gpu
When running with multi-gpu, excessive tracing is caused. To allviate this, we
can run without eager mode and trace a single graph, but this then causes
errors such as:
    316     NotImplementedError: Cannot convert a symbolic Tensor
    (cond_2/Identity_1:0) to a numpy array. This error may indicate that you're
    trying to pass a Tensor to a NumPy call, which is not supported

Although this is slow, perhaps if epochs limited to 40, or if cpu time exceeds
x num of hours it should allow for model to be trained to completion, just
without full potential epochs to convergence.

	modified:   astronet/atx/train.py
	modified:   bin/plasticc-myriad