Add 'scaledown_factor' to reduce overfitting
It seems the many parameters for the original model are causing
overfitting to the training data.

By reducing the number of filters, but retaining the over architecture,
this should hopefully improve the validation scores.

These should be called in the first instance with:

    $ # python $ASNWD/astronet/$ARCH/train.py --dataset $dataset --epochs 400 -z "" -t "" --model "scaledown-by-X"

Where X is either, 2, 4, 8. A true optimisation should then be done
later after trial runs.

	modified:   astronet/atx/layers.py
	modified:   astronet/atx/model.py
	modified:   astronet/atx/opt/hypertrain.py
	modified:   astronet/atx/opt/runs/plasticc/results_with_z.json