Use robust scaler on numpy array instead of df
There is unnecessary overhead applying to dataframe when this can be
directly run on the numpy array; results in a 2x speed up.

Although robust scaling was applied in previous studies, it has not been
applied to the "full test set" of PLAsTiCC. This was initially because
of computational resources, but it should also be noted that there is a
chance of data leakage in some cases where it is  applied _before_ train
and test split.

Refs:
  - https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.robust_scale.html#sklearn.preprocessing.robust_scale

  Warning Risk of data leak
  Do not use robust_scale unless you know what you are doing. A common
  mistake is to apply it to the entire data before splitting into training
  and test sets. This will bias the model evaluation because information
  would have leaked from the test set to the training set. In general, we
  recommend using RobustScaler within a Pipeline in order to prevent most
  risks of data leaking: pipe = make_pipeline(RobustScaler(),
  LogisticRegression()).

	new file:   t2-mwe-ztf-compressed-clustered-tflite-model-rs.lnprofile
	modified:   t2-mwe-ztf.py
	modified:   t2-mwe-ztf.py.lprof
	modified:   t2-ztf-logloss-check.py